@@ -3,14 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "dompurify": "^2.3.3",
-    "marked": "^3.0.8",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "sass": "^1.43.4",
-    "typescript": "^4.4.4",
+    "dompurify": "^2.3.8",
+    "marked": "^4.0.17",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^5.3.3",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.53.0",
+    "typescript": "^4.7.4",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
@@ -39,22 +39,22 @@
     ]
   },
   "devDependencies": {
-    "@apollo/client": "^3.5.6",
-    "@graphql-codegen/cli": "^2.2.2",
-    "@graphql-codegen/schema-ast": "^2.4.0",
-    "@graphql-codegen/typescript": "^2.3.0",
-    "@graphql-codegen/typescript-operations": "^2.2.1",
-    "@graphql-codegen/typescript-react-apollo": "^3.2.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/dompurify": "^2.3.1",
-    "@types/jest": "^27.0.2",
-    "@types/marked": "^3.0.2",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-router-dom": "^5.3.2",
-    "graphql": "^15.7.2"
+    "@apollo/client": "^3.6.9",
+    "@graphql-codegen/cli": "^2.7.0",
+    "@graphql-codegen/schema-ast": "^2.4.1",
+    "@graphql-codegen/typescript": "^2.6.0",
+    "@graphql-codegen/typescript-operations": "^2.4.3",
+    "@graphql-codegen/typescript-react-apollo": "^3.2.17",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
+    "@types/dompurify": "^2.3.3",
+    "@types/jest": "^28.1.4",
+    "@types/marked": "^4.0.3",
+    "@types/node": "^18.0.1",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "^5.3.3",
+    "graphql": "^16.5.0"
   }
 }
