@@ -1,5 +1,5 @@
 import DOMPurify from 'dompurify';
-import marked from 'marked';
+import { marked } from 'marked';
 import { useEffect, useState } from 'react';
 import { slugify } from '../../helpers';
 
